requirements.txt
# Unified Requirements for TripMind-AI Development & Deployment

# Backend Dependencies
langchain==0.2.14
langchain-core==0.2.34
langgraph==0.2.14
langchain-google-genai==1.0.8
langchain-huggingface==0.0.3
tavily-python==0.3.3
duckduckgo-search==6.2.13
langchain-community==0.2.12
faiss-cpu==1.8.0
pypdf==4.3.1
langsmith==0.1.104
python-dotenv==1.0.1
langchain-groq==0.1.9
langchain-openai==0.1.22

# Frontend Dependencies
streamlit==1.37.1
requests==2.32.3
